requirements.txt
python-dotenv
supabase
notion-client
openai
markdown